# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Trade-J API
  version: 1.0.0
  description: |
    Java 21 algorithmic trading platform — live market data, OMS, risk, strategy plugins,
    and broker adapters (Dhan, Upstox, ICICI Direct).
  contact:
    name: Trade-J Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.tradej.example.com
    description: Production
paths:
...
  /api/v1/market/ltp:
    get:
      tags: [Market Data]
      summary: Last Traded Price
      parameters:
        - name: symbol
          in: query
          required: true
          schema: { type: string }
          example: SBIN
        - name: exchangeSegment
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ExchangeSegment'
      responses:
        '200':
          description: LTP response
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbol: { type: string }
                  exchangeSegment: { type: string }
                  ltpPaisa: { type: integer }
                  exchangeTimestampMs: { type: integer }
  /api/v1/market/historical/candles:
    get:
      tags: [Market Data]
      summary: Historical Candles
      parameters:
        - name: symbol
          in: query
          required: true
          schema: { type: string }
        - name: exchangeSegment
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ExchangeSegment'
        - name: interval
          in: query
          schema:
            type: string
            default: 1d
          description: Candle interval (e.g. 1m, 5m, 15m, 1h, 1d)
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: source
          in: query
          schema:
            type: string
            default: broker
            enum: [broker, parquet]
      responses:
        '200':
          description: Candle array
...
components:
  schemas:
    ExchangeSegment:
      type: string
      enum: [NSE_EQ, NSE_FNO, BSE_EQ, BSE_FNO, MCX_COMM, IDX_I, NSE_CURRENCY, BSE_CURRENCY]
...